chore: add tests, docs. export the compile method
