+++ .gitignore
@@ -2,6 +2,7 @@
 
 .moon/cache
 .eslintcache
+.compiled
 .rendered
 .test
 node_modules
