+++ docs/recipes.md
@@ -7,13 +7,23 @@ description: 'Recipes for jsx-email'
 
 Recipes are actual, working examples of jsx-email features and techniques that can be copied and pasted. They help users to understand how a thing works outside of the documentation.
 
-Available recipes:
+### Available recipes:
 
 <table border="0" cellspacing="0" cellpadding="0" class="recipes">
  <tr>
     <td>🧁</td>
     <td>Importing CSS</td>
     <td width="100%"><a href="https://github.com/shellscape/jsx-email/tree/main/recipes/import-css">/recipes/import-css</a></td>
+ </tr>
+</table>
+
+### External Recipes
+
+<table border="0" cellspacing="0" cellpadding="0" class="recipes">
+ <tr>
+    <td>🧁</td>
+    <td>Integration with Keycloak(ify)</td>
+    <td width="100%"><a href="https://github.com/timofei-iatsenko/keycloakify-jsx-emails">github.com/timofei-iatsenko/keycloakify-jsx-emails</a></td>
  </tr>
 </table>
