+++ packages/jsx-email/src/helpers.ts
@@ -1,0 +1,2 @@
+// Note: after server start we change the root directory to trick vite
+export const originalCwd = process.cwd();
