+++ packages/jsx-email/test/render/fixtures/async-template.tsx
@@ -1,4 +1,4 @@
-import { Suspense } from 'react';
+import React, { Suspense } from 'react';
 
 import { useData } from '../../../src/renderer/suspense.js';
