+++ packages/jsx-email/test/render/fixtures/import-css.css
@@ -1,0 +1,3 @@
+body {
+  background: #000;
+}
