+++ packages/jsx-email/test/render/fixtures/import-css.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import css from './import-css.css';
+
+export const Template: React.FC = () => (
+  <>
+    <style>{css as string}</style>
+  </>
+);
