+++ recipes/import-css/README.md
@@ -13,3 +13,11 @@ $ npm i && npm run dev
 ```
 
 Once the preview app opens, select the `Import CSS` template and examine the HTML and JSX code tabs to view the source and result of a CSS import.
+
+## Notes
+
+_Caveat: Importing CSS is only supported when using the CLI [`build`](https://jsx.email/docs/core/cli#build) or [`preview`](https://jsx.email/docs/core/cli#preview) commands, or the [`compile`](https://jsx.email/docs/core/render) API method_
+
+It's important to remember that jsx-email templates _are not_ React apps. This is something that often trips up users; where they expect a React app feature, such as importing CSS, to work exactly like a React app. With React, the user is _always_ compiling the app to run it, and React's bundler configurations provide for things like automatic import and injection of CSS.
+
+The `render` method simply renders JSX/TSX; it doesn't perform a bundle or a build, hence traversing imports isn't something that it does. If a use case calls for using the API instead of the CLI and importing CSS is needed, use the `compile` API first, import the `Template` from the compiled file, and pass the `Template` to `render`.
